requirements.txt
Flask==3.1.0
tensorflow==2.10.1
opencv-python==4.11.0.86
numpy==1.26.4
pandas==2.2.3
requests==2.32.3
google-auth==2.38.0
google-auth-oauthlib==0.4.6
protobuf==3.19.6
Werkzeug==3.1.3
Jinja2==3.1.6
itsdangerous==2.2.0